requirements.txt
scikit_image==0.18.1
yacs==0.1.8
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tqdm==4.59.0
opencv_python==4.5.3.56
pycocotools==2.0.2
numpy==1.20.1
detectron2.egg==info
matplotlib==3.3.4
tabulate==0.8.9
fvcore==0.1.5.post20210630
cityscapesscripts==2.2.0
ipython==7.27.0
mysql_connector_repackaged==0.3.1
Pillow==8.3.1
PyQt5==5.15.4
pyserial==3.5
skimage==0.0
